requirements.txt
PyYAML==5.1.2
slackclient==2.1.0
